add drawio file; add debug output in code

+++ src/meshfem3D/meshfem3D_par.f90
@@ -243,7 +243,7 @@
   implicit none
 
   ! topology of the elements (for corners)
-  integer, dimension(NGNOD) :: iaddx_corner,iaddy_corner,iaddz_corner
+  double precision, dimension(NGNOD) :: iaddx_corner,iaddy_corner,iaddz_corner
 
   ! Gauss-Lobatto-Legendre points and weights of integration
   double precision, dimension(NGLLX) :: xigll,wxgll
